:recycle: refactor : 라우팅 경로에 따른 Link text color 변경
